requirements.txt
mcp[cli]>=1.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
langgraph>=0.2.0
langchain-ollama>=0.1.0
langchain-core>=0.3.0
langchain-community>=0.3.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0